Refactoring ContrabandTable
