+++ traffic_stops/static/js/app/states/nc/ContrabandHitRate.js
@@ -18,7 +18,7 @@ const ContrabandHitRateHandler = C.ContrabandHitRateHandlerBase.extend({
 
 // dashboard visuals
 
-var ContrabandHitRateBar = C.ContrabandHitRateBarBase.extend({
+const ContrabandHitRateBar = C.ContrabandHitRateBarBase.extend({
   defaults: {
     showEthnicity: true,
     width: 750,
@@ -40,34 +40,12 @@ var ContrabandHitRateBar = C.ContrabandHitRateBarBase.extend({
   }
 });
 
-var ContrabandTable = TableBase.extend({
-  get_tabular_data: function(){
-    var header, row, rows = [], se, cb;
-
-    // create header
-    header = ["Year"];
-    header.push.apply(header, Stops.pprint.values());
-    rows.push(header);
-
-    var raw = this.data.raw,
-        searches = _.object(_.pluck(raw.searches, 'year'), raw.searches),
-        contrabands = _.object(_.pluck(raw.contraband, 'year'), raw.contraband);
-
-    _.keys(searches).forEach(function(yr){
-      se = searches[yr];
-      cb = contrabands[yr] || {};
-      row = [yr];
-      Stops.races.forEach(function(r){
-        row.push((cb[r]||0).toLocaleString() + "/" + (se[r]||0).toLocaleString());
-      });
-      Stops.ethnicities.forEach(function(e){
-        row.push((cb[e]||0).toLocaleString() + "/" + (se[e]||0).toLocaleString());
-      });
-      rows.push(row);
-    });
-
-    return rows;
+const ContrabandTable = C.ContrabandTableBase.extend({
+  _get_header_rows: function () {
+    return Stops.pprint.values();
   }
+
+  types: [Stops.races, Stops.ethnicities]
 });
 
 export default {
